requirements.txt
about-time==4.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alive-progress==3.1.0
altgraph==0.17.3
annotated-types==0.5.0
anyio==3.6.2
apikey==0.2.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.15.2
asttokens==2.2.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.2.0
auto-py-to-exe==2.30.0
backcall==0.2.0
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blessings==1.7
blinker==1.6.2
blis==0.7.9
bottle==0.12.24
bottle-websocket==0.2.9
cachelib==0.10.2
captcha==0.5.0
catalogue==2.0.8
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
commonmark==0.9.1
comtypes==1.2.0
confection==0.0.4
contourpy==1.0.7
copier==8.3.0
credentials==1.1
Crypto.Cipher==1
customtkinter==5.1.2
cvxpy==1.3.1
cycler==0.11.0
cymem==2.0.7
Cython==0.29.36
darkdetect==0.8.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.17.1
dill==0.3.6
dunamai==1.18.0
ecos==2.0.12
Eel==0.14.0
executing==1.2.0
fake-useragent==1.1.3
fastapi==0.88.0
fastjsonschema==2.17.1
filelock==3.12.2
Flask==2.3.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-Session==0.5.0
flask-session-captcha==1.3.0
Flask-Sessionstore==0.4.5
flask-simplelogin==0.1.2
Flask-SocketIO==5.3.5
Flask-SQLAlchemy==3.0.3
Flask-SSE==1.0.0
Flask-WTF==1.1.1
flet==0.10.1
flet-core==0.10.1
flet-runtime==0.10.1
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
ftransc==7.0.3
funcy==2.0
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
grapheme==0.6.0
greenlet==2.0.2
gTTS==2.3.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
install==1.3.5
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso639==0.1.4
isoduration==20.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.9.0
lazy_loader==0.2
littleutils==0.2.2
mapping==0.1.6
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mutagen==1.47.0
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
opencv-python==4.7.0.72
osqp==0.6.2.post9
outdated==0.2.2
packaging==23.1
pafy==0.5.5
pandas==1.5.3
pandas-flavor==0.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pathy==0.10.1
patsy==0.5.3
pefile==2023.2.7
pickleshare==0.7.5
Pillow==9.4.0
pingouin==0.5.3
platformdirs==3.2.0
plotly==5.13.1
plumbum==1.8.2
polars==0.16.17
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py2puml==0.7.2
pyasn1==0.4.8
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.17
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pygame==2.2.0
pyglet==1.5.27
Pygments==2.14.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyJWT==2.8.0
pylint==2.17.2
pynecone==0.1.18
pyparsing==3.0.9
pypiwin32==223
pypng==0.20220715.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.0
pyrsistent==0.19.3
PySide6==6.5.0
PySide6-Addons==6.5.0
PySide6-Essentials==6.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-json-logger==2.0.7
python-slugify==8.0.1
python-socketio==5.7.2
pyttsx3==2.90
pytube==15.0.0
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyyaml-include==1.3.1
pyzmq==25.0.2
qdldl==0.1.7
qrcode==7.4.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
qtconsole==5.4.3
QtPy==2.3.1
questionary==2.0.1
redis==4.5.1
regex==2023.6.3
repath==0.9.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
screen-brightness-control==0.22.1
scs==3.2.3
seaborn==0.12.2
Send2Trash==1.8.2
shiboken6==6.5.0
six==1.16.0
sketchpy==0.1.0
smart-open==6.3.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.16
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
srsly==2.4.6
stack-data==0.6.2
starlette==0.22.0
statsmodels==0.14.0
svg.path==6.2
svgpathtools==1.6.0
svgwrite==1.4.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
textblob==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tomlkit==0.11.7
torch==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
ttkbootstrap==1.10.1
turtle==0.0.1
typer==0.4.2
typing==3.7.4.3
typing_extensions==4.8.0
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
waitress==2.1.2
wasabi==1.1.1
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.0
whichcraft==0.6.1
widgetsnbextension==4.0.7
wincertstore==0.2
windows-curses==2.3.1
WMI==1.5.1
wrapt==1.15.0
WTForms==3.0.1
xarray==2023.5.0
yarl==1.8.2
youtube-dl==2021.12.17
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
